apps/api/requirements.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Scraper dependencies
beautifulsoup4>=4.9.0
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
lxml>=4.6.0
python-dateutil>=2.8.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov>=2.12.0
pytest-mock>=3.6.0
factory-boy>=3.2.0